requirements.txt
Django==1.6.5
MySQL-python==1.2.5
South==1.0
django-cors-headers==0.13
djangorestframework==2.3.14
pymongo==2.7.2
wsgiref==0.1.2
